fix: naming karmoisin -> karmosin
