+++ src/main/java/karmosin/ContinuousIntegrationJobTaskOutput.java
@@ -1,4 +1,4 @@
-package karmoisin;
+package karmosin;
 
 public class ContinuousIntegrationJobTaskOutput {
